+++ CHANGELOG.md
@@ -8,7 +8,8 @@ Include the issue number (#xxx) which will link back to the originating issue
 in github. Commentary on the change should appear as a nested, unordered list.
 
 ## 1.0.11 (WIP)
-
+- Improvements
+  - Add the --custom-report option to build your own formatter
 - Bugfixes
   - Fix the handling of --ns-regex and --test-ns-regex (#183)
   - Read options from the Leiningen project file (#155)